playground/py_display_simulator/requirements.txt
PyQt5
pyserial
Pillow
